add MemoryAllocator/MemoryBlock codes.
